Move URLRequest extension and PlaceholderDataValidator to separate files
