+++ Networking/Networking/Extensions/URLRequest+Request.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+/// Makes URLRequest conform to Request.
+extension URLRequest: Request {
+    func responseDataValidator() -> ResponseDataValidator {
+        PlaceholderDataValidator()
+    }
+}
