+++ Networking/Networking/Validators/PlaceholderDataValidator.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+/// Placeholder implementation for `ResponseDataValidator`.
+///
+final class PlaceholderDataValidator: ResponseDataValidator {
+    func validate(data: Data) throws {
+        // no-op
+    }
+}
